- Remove unnecessary interface
