--- feeder-web/src/main/kotlin/ru/timmson/feeder/bot/BotServiceImpl.kt
@@ -1,44 +1,0 @@
-package ru.timmson.feeder.bot
-
-import com.pengrad.telegrambot.request.SendMessage
-import jakarta.annotation.PostConstruct
-import jakarta.annotation.PreDestroy
-import org.springframework.stereotype.Service
-import ru.timmson.feeder.common.FeederConfig
-import ru.timmson.feeder.common.logger
-import java.time.LocalDateTime
-
-
-@Service
-class BotServiceImpl(
-    private val bot: BotProxy,
-    private val feederConfig: FeederConfig,
-    private val botListener: BotListener
-) : BotService {
-
-    private val log = logger<BotServiceImpl>()
-
-    @PostConstruct
-    fun postConstruct() {
-        bot.startup(feederConfig.token)
-        bot.setUpdatesListener(botListener)
-        sendMessageToOwner("Started at ${LocalDateTime.now()}")
-    }
-
-    override fun sendMessage(chatId: Any, messageText: String) {
-        log.info("Entering sendMessage ($chatId, '$messageText') ...")
-        val response = bot.execute(SendMessage(chatId, messageText))
-        log.info("Leaving sendMessage (...) = ${response.isOk}")
-    }
-
-    override fun sendMessageToOwner(messageText: String) {
-        sendMessage(feederConfig.ownerId, messageText)
-    }
-
-    @PreDestroy
-    fun preDestroy() {
-        bot.removeGetUpdatesListener()
-        bot.shutdown()
-    }
-
-}
