+++ feeder-web/src/test/kotlin/ru/timmson/feeder/bot/BotProxyShould.kt
@@ -8,13 +8,13 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.junit.jupiter.MockitoExtension
 
 @ExtendWith(MockitoExtension::class)
-class BotProxyImplShould {
+class BotProxyShould {
 
-    private lateinit var botProxy: BotProxyImpl
+    private lateinit var botProxy: BotProxy
 
     @BeforeEach
     fun setUp() {
-        botProxy = BotProxyImpl()
+        botProxy = BotProxy()
     }
 
     @Test
